+++ app/src/main/java/com/ceaver/assin/markets/Title.kt
@@ -5,6 +5,7 @@ import androidx.room.Entity
 import androidx.room.Index
 import androidx.room.PrimaryKey
 import com.ceaver.assin.assets.AssetCategory
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 @Entity(tableName = "title", indices = [Index(value = ["symbol", "rank"])]) // TODO What is this index good for? https://developer.android.com/training/data-storage/room/defining-data
@@ -23,44 +24,44 @@ data class Title(//
         @ColumnInfo(name = "betaValue") val betaValue: Double? = null,
         @ColumnInfo(name = "lastUpdated") val lastUpdated: LocalDateTime? = null,
         // usd
-        @ColumnInfo(name = "priceUsd") val priceUsd: Double? = null,
-        @ColumnInfo(name = "volume24hUsd") val volume24hUsd: Double? = null,
-        @ColumnInfo(name = "marketCapUsd") val marketCapUsd: Double? = null,
-        @ColumnInfo(name = "marketCapChange24hUsd") val marketCapChange24hUsd: Double? = null,
-        @ColumnInfo(name = "percentChange1hUsd") val percentChange1hUsd: Double? = null,
-        @ColumnInfo(name = "percentChange24hUsd") val percentChange24hUsd: Double? = null,
-        @ColumnInfo(name = "percentChange7dUsd") val percentChange7dUsd: Double? = null,
-        @ColumnInfo(name = "percentChange30dUsd") val percentChange30dUsd: Double? = null,
-        @ColumnInfo(name = "percentChange1yUsd") val percentChange1yUsd: Double? = null,
-        @ColumnInfo(name = "athPriceUsd") val athPriceUsd: Double? = null,
+        @ColumnInfo(name = "priceUsd") val priceUsd: BigDecimal? = null,
+        @ColumnInfo(name = "volume24hUsd") val volume24hUsd: BigDecimal? = null,
+        @ColumnInfo(name = "marketCapUsd") val marketCapUsd: BigDecimal? = null,
+        @ColumnInfo(name = "marketCapChange24hUsd") val marketCapChange24hUsd: BigDecimal? = null,
+        @ColumnInfo(name = "percentChange1hUsd") val percentChange1hUsd: BigDecimal? = null,
+        @ColumnInfo(name = "percentChange24hUsd") val percentChange24hUsd: BigDecimal? = null,
+        @ColumnInfo(name = "percentChange7dUsd") val percentChange7dUsd: BigDecimal? = null,
+        @ColumnInfo(name = "percentChange30dUsd") val percentChange30dUsd: BigDecimal? = null,
+        @ColumnInfo(name = "percentChange1yUsd") val percentChange1yUsd: BigDecimal? = null,
+        @ColumnInfo(name = "athPriceUsd") val athPriceUsd: BigDecimal? = null,
         @ColumnInfo(name = "athDateUsd") val athDateUsd: LocalDateTime? = null,
-        @ColumnInfo(name = "athPercentUsd") val athPercentUsd: Double? = null,
+        @ColumnInfo(name = "athPercentUsd") val athPercentUsd: BigDecimal? = null,
         // btc
-        @ColumnInfo(name = "priceBtc") val priceBtc: Double? = null,
-        @ColumnInfo(name = "volume24hBtc") val volume24hBtc: Double? = null,
-        @ColumnInfo(name = "marketCapBtc") val marketCapBtc: Double? = null,
-        @ColumnInfo(name = "marketCapChange24hBtc") val marketCapChange24hBtc: Double? = null,
-        @ColumnInfo(name = "percentChange1hBtc") val percentChange1hBtc: Double? = null,
-        @ColumnInfo(name = "percentChange24hBtc") val percentChange24hBtc: Double? = null,
-        @ColumnInfo(name = "percentChange7dBtc") val percentChange7dBtc: Double? = null,
-        @ColumnInfo(name = "percentChange30dBtc") val percentChange30dBtc: Double? = null,
-        @ColumnInfo(name = "percentChange1yBtc") val percentChange1yBtc: Double? = null,
-        @ColumnInfo(name = "athPriceBtc") val athPriceBtc: Double? = null,
+        @ColumnInfo(name = "priceBtc") val priceBtc: BigDecimal? = null,
+        @ColumnInfo(name = "volume24hBtc") val volume24hBtc: BigDecimal? = null,
+        @ColumnInfo(name = "marketCapBtc") val marketCapBtc: BigDecimal? = null,
+        @ColumnInfo(name = "marketCapChange24hBtc") val marketCapChange24hBtc: BigDecimal? = null,
+        @ColumnInfo(name = "percentChange1hBtc") val percentChange1hBtc: BigDecimal? = null,
+        @ColumnInfo(name = "percentChange24hBtc") val percentChange24hBtc: BigDecimal? = null,
+        @ColumnInfo(name = "percentChange7dBtc") val percentChange7dBtc: BigDecimal? = null,
+        @ColumnInfo(name = "percentChange30dBtc") val percentChange30dBtc: BigDecimal? = null,
+        @ColumnInfo(name = "percentChange1yBtc") val percentChange1yBtc: BigDecimal? = null,
+        @ColumnInfo(name = "athPriceBtc") val athPriceBtc: BigDecimal? = null,
         @ColumnInfo(name = "athDateBtc") val athDateBtc: LocalDateTime? = null,
-        @ColumnInfo(name = "athPercentBtc") val athPercentBtc: Double? = null,
+        @ColumnInfo(name = "athPercentBtc") val athPercentBtc: BigDecimal? = null,
         // eth
-        @ColumnInfo(name = "priceEth") val priceEth: Double? = null,
-        @ColumnInfo(name = "volume24hEth") val volume24hEth: Double? = null,
-        @ColumnInfo(name = "marketCapEth") val marketCapEth: Double? = null,
-        @ColumnInfo(name = "marketCapChange24hEth") val marketCapChange24hEth: Double? = null,
-        @ColumnInfo(name = "percentChange1hEth") val percentChange1hEth: Double? = null,
-        @ColumnInfo(name = "percentChange24hEth") val percentChange24hEth: Double? = null,
-        @ColumnInfo(name = "percentChange7dEth") val percentChange7dEth: Double? = null,
-        @ColumnInfo(name = "percentChange30dEth") val percentChange30dEth: Double? = null,
-        @ColumnInfo(name = "percentChange1yEth") val percentChange1yEth: Double? = null,
-        @ColumnInfo(name = "athPriceEth") val athPriceEth: Double? = null,
+        @ColumnInfo(name = "priceEth") val priceEth: BigDecimal? = null,
+        @ColumnInfo(name = "volume24hEth") val volume24hEth: BigDecimal? = null,
+        @ColumnInfo(name = "marketCapEth") val marketCapEth: BigDecimal? = null,
+        @ColumnInfo(name = "marketCapChange24hEth") val marketCapChange24hEth: BigDecimal? = null,
+        @ColumnInfo(name = "percentChange1hEth") val percentChange1hEth: BigDecimal? = null,
+        @ColumnInfo(name = "percentChange24hEth") val percentChange24hEth: BigDecimal? = null,
+        @ColumnInfo(name = "percentChange7dEth") val percentChange7dEth: BigDecimal? = null,
+        @ColumnInfo(name = "percentChange30dEth") val percentChange30dEth: BigDecimal? = null,
+        @ColumnInfo(name = "percentChange1yEth") val percentChange1yEth: BigDecimal? = null,
+        @ColumnInfo(name = "athPriceEth") val athPriceEth: BigDecimal? = null,
         @ColumnInfo(name = "athDateEth") val athDateEth: LocalDateTime? = null,
-        @ColumnInfo(name = "athPercentEth") val athPercentEth: Double? = null
+        @ColumnInfo(name = "athPercentEth") val athPercentEth: BigDecimal? = null
 ) {
 
     fun inactive(): Boolean {
